requirements.txt
cryptography==41.0.5
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.3
django-money==3.4.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
Pillow==10.1.0
PyJWT==2.8.0
